refactor(webui): extract Snapshot Binder (entities/renderer/snapshot/)
useRendererOutput.ts:23-55의 변환 helper(toRendererAgentState,
reuseStableFiles)와 snapshot 조립 로직을 entities/renderer/snapshot/
의 순수 함수로 추출한다.

- toRendererAgentState: AgentState -> RendererAgentState.
- reuseStableFiles: digest/modifiedAt 기반 파일 재사용.
- buildRendererSnapshot: slug + agent state + files -> RendererSnapshot
  조립.

useRendererOutput hook은 이제 fetch + 변환 호출 + dispatch만 한다.

issue #144 Phase F2.
