+++ apps/webui/src/client/entities/renderer/snapshot/index.ts
@@ -1,0 +1,5 @@
+export {
+  buildRendererSnapshot,
+  reuseStableFiles,
+  toRendererAgentState,
+} from "./toSnapshot.js";
